Create Balanced_Delimiters
You are given a text that contains multiple types of delimiters that act like parentheses. Implement a program that checks if all delimiteres in the input text are properly balanced.

Requirements. Your solution must provide an implementation of LinkedStack (implementation based on linked lists) and use it to solve the problem. You cannot use any standard data structures such as ArrayList and LinkedList in Java or std::vector and std::list in C++.