Create Accounting_for_a_café
A local café would like to keep track of its earnings. Your task is to implement a simple accounting program for the café. The program is given a sequence of entries, one per sold item. Each entry contains a timestamp, receipt number, sold item title and its cost. The task is to compute for each date (year, month and day) total amount of money earned, as well as number of unique customers (number of unique receipts).

Requirements. You must provide an implementation of HashMap and use it to solve the problem. You must implement a rehashing strategy (dynamic hashtable) to keep the load factor constant. You may use primitive arrays, as well as ArrayList, LinkedList in Java and std::vector, std::list in C++. The tests must demonstrate that your implementation of rehashing works properly. Thus, the initial size of the hashtable must be sufficiently small (≤100
).